Initial Pester Tests (#19)
* WaykNow-ps: add Get-WaykNowPath function to help stubbing tests

* WaykNow-ps: use CmdletBinding declaration on exported functions

* WaykNow-ps: minor formatting

* WaykNow-ps: correctly mock Get-WaykNowPath

* WaykNow-ps: fix cases where directory and files do not exist prior to running the script

+++ WaykNow/Public/WaykNowLicense.ps1
@@ -1,13 +1,23 @@
 function Set-WaykNowLicense
-(    
-    [Parameter(Mandatory = $true, HelpMessage= "Wayk Now License")]
-    [string] $License
-)
 {
+    [CmdletBinding()]
+    param(
+        [Parameter(Mandatory,Position=0)]
+        [string] $License
+    )
+
     $licensePattern = '[A-Z0-9]{5}-[A-Z0-9]{5}-[A-Z0-9]{5}-[A-Z0-9]{5}-[A-Z0-9]{5}'
     $WaykNowInfo = Get-WaykNowInfo
+    $ConfigFile = $WaykNowInfo.ConfigFile
+
     if ($License -CMatch $licensePattern) {
-        $json = Get-Content -Path $WaykNowInfo.ConfigFile -Raw -Encoding UTF8 | ConvertFrom-Json
+
+        if (Test-Path $ConfigFile) {
+            $json = Get-Content -Path $ConfigFile -Raw -Encoding UTF8 | ConvertFrom-Json
+        } else {
+            $json = '{}' | ConvertFrom-Json
+        }
+
         if ($json.RegistrationSerial)
         {
             $json.RegistrationSerial = $License;
@@ -15,7 +25,7 @@ function Set-WaykNowLicense
         else
         {
             # If the json is empty
-            if(!$json){
+            if (!$json) {
                 $json = '{}'
                 $json = ConvertFrom-Json $json
             }
@@ -23,31 +33,47 @@ function Set-WaykNowLicense
             $json | Add-Member -Type NoteProperty -Name 'RegistrationSerial' -Value $License -Force
         }
 
+        New-Item -Path $(Split-Path $ConfigFile -Parent) -ItemType 'Directory' -Force
+
         $fileValue = $json | ConvertTo-Json
         $Utf8NoBomEncoding = New-Object System.Text.UTF8Encoding $False
-        [System.IO.File]::WriteAllLines($WaykNowInfo.ConfigFile, $fileValue, $Utf8NoBomEncoding)
-    }
-    else{
+        [System.IO.File]::WriteAllLines($ConfigFile, $fileValue, $Utf8NoBomEncoding)
+    } else {
         Write-Error "Invalid License Format"
     }
 }
 
-function Get-WaykNowLicense {
-    [WaykNowInfo]$WaykInfo = Get-WaykNowInfo
-    $json = Get-Content -Path $WaykInfo.ConfigFile -Raw -Encoding UTF8 | ConvertFrom-Json
+function Get-WaykNowLicense
+{
+    [CmdletBinding()]
+    param()
 
-    return $json.RegistrationSerial
+    [WaykNowInfo]$WaykInfo = Get-WaykNowInfo
+
+    if (Test-Path $WaykInfo.ConfigFile) {
+        $json = Get-Content -Path $WaykInfo.ConfigFile -Raw -Encoding UTF8 | ConvertFrom-Json
+        return $json.RegistrationSerial;
+    } else {
+        return $null
+    }
 }
 
-function Reset-WaykNowLicense {
+function Reset-WaykNowLicense
+{
+    [CmdletBinding()]
+    param()
+
     [WaykNowInfo]$WaykInfo = Get-WaykNowInfo
 
-    $json = Get-Content -Path $WaykInfo.ConfigFile -Raw -Encoding UTF8 | ConvertFrom-Json
-    if ($json.RegistrationSerial) {
-        $json.RegistrationSerial = ''
-        $fileValue = $json | ConvertTo-Json
-        $Utf8NoBomEncoding = New-Object System.Text.UTF8Encoding $False
-        [System.IO.File]::WriteAllLines($WaykInfo.ConfigFile, $fileValue, $Utf8NoBomEncoding)
+    if (Test-Path $WaykInfo.ConfigFile) {
+        $json = Get-Content -Path $WaykInfo.ConfigFile -Raw -Encoding UTF8 | ConvertFrom-Json
+
+        if ($json.RegistrationSerial) {
+            $json.RegistrationSerial = ''
+            $fileValue = $json | ConvertTo-Json
+            $Utf8NoBomEncoding = New-Object System.Text.UTF8Encoding $False
+            [System.IO.File]::WriteAllLines($WaykInfo.ConfigFile, $fileValue, $Utf8NoBomEncoding)
+        }
     }
 }
